% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_data.R
\name{get_value_data}
\alias{get_value_data}
\title{Get image values}
\usage{
get_value_data(img1, img2, mask = NULL)
}
\arguments{
\item{img1}{Object of class \code{nifti}, array, or filename of image}

\item{img2}{Object of class \code{nifti}, array, or filename of image}

\item{mask}{Binary bject of class \code{nifti}, array, or filename of image.
If \code{NULL}}
}
\value{
List of values from the images
}
\description{
Simple wrapper for robust extraction of
values from images
}

