% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_entropy.R
\name{joint_entropy}
\alias{joint_entropy}
\title{Joint Entropy}
\usage{
joint_entropy(joint)
}
\arguments{
\item{joint}{Matrix of joint densities.}
}
\value{
Scalar value of entropy
}
\description{
Calculates Shannon entropy from a joint density.
}

