% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_to_mi.R
\name{joint_to_mi}
\alias{joint_to_mi}
\title{Joint PDF to mutual information}
\usage{
joint_to_mi(joint)
}
\arguments{
\item{joint}{Matrix of joint densities.  Does not need to be square.}
}
\value{
Scalar value of mutual information
}
\description{
Calculates mutual information from a joint density.
}

