% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde2d.nifti.R
\name{kde2d.nifti}
\alias{kde2d.nifti}
\title{Two-Dimensional Kernel Density Estimation of Images}
\usage{
kde2d.nifti(img1, img2, mask = NULL, ...)
}
\arguments{
\item{img1}{Object of class \code{nifti}, array, or filename of image}

\item{img2}{Object of class \code{nifti}, array, or filename of image}

\item{mask}{Binary bject of class \code{nifti}, array, or filename of image.
If \code{NULL}}

\item{...}{Additional arguments to pass to \code{\link{kde2d}}}
}
\value{
Matrix of the joint PDF
}
\description{
Two-dimensional kernel density estimation of
2 nifti images
}

